+++ src/gennylib/src/context.cpp
@@ -84,7 +84,7 @@ ActorVector WorkloadContext::_constructActors(const Cast& cast,
     return actors;
 }
 
-mongocxx::pool::entry WorkloadContext::client(const std::string& name, unsigned long instance) {
+mongocxx::pool::entry WorkloadContext::client(const std::string& name, size_t instance) {
     return _poolManager.client(name, instance, *this);
 }
